delete some more unused and refactor files
