+++ src/components/CategorizationRules/CategorizationRuleForm/CategoryMobileDrawer.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useMemo } from 'react'
 
-import { getLeafCategories } from '@internal-types/categories'
 import { CategoryAsOption } from '@internal-types/categorizationOption'
 import { type Classification, ClassificationEquivalence } from '@schemas/categorization'
+import { getLeafCategories } from '@utils/categories'
 import { useCategories } from '@hooks/api/businesses/[business-id]/categories/useCategories'
 import type { BankTransactionNonSuggestedMatchOption } from '@providers/BankTransactionsCategorizationStore/utils'
 import { VStack } from '@ui/Stack/Stack'
