+++ src/components/CategorizationRules/CategorizationRulesView/ResponsiveCategorizationRulesView.tsx
@@ -2,9 +2,9 @@ import { useCallback, useMemo, useState } from 'react'
 import { PencilRuler, Plus } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 
-import { flattenCategories } from '@internal-types/categories'
 import type { CategorizationRule } from '@schemas/bankTransactions/categorizationRules/categorizationRule'
 import { CategoriesListMode } from '@schemas/categorization'
+import { flattenCategories } from '@utils/categories'
 import { BREAKPOINTS } from '@utils/screenSizeBreakpoints'
 import { useCategories } from '@hooks/api/businesses/[business-id]/categories/useCategories'
 import { useArchiveCategorizationRule } from '@hooks/api/businesses/[business-id]/categorization-rules/[categorization-rule-id]/archive/useArchiveCategorizationRule'
