+++ src/components/CategorizationRules/ResolvedCategoryName.tsx
@@ -1,6 +1,6 @@
-import { getResolvedCategoryName } from '@internal-types/categories'
 import type { AccountIdentifier } from '@schemas/accountIdentifier'
 import type { NestedCategorization } from '@schemas/categorization'
+import { getResolvedCategoryName } from '@utils/categories'
 import { Span, type TextStyleProps } from '@ui/Typography/Text'
 
 type ResolvedCategoryNameProps = {
