+++ src/utils/categories.ts
@@ -1,36 +1,5 @@
 import { type AccountIdentifier, AccountIdEquivalence, AccountStableNameEquivalence, makeAccountId, makeStableName } from '@schemas/accountIdentifier'
-import { type CategorizationEncoded as ApiCategorization, type NestedCategorization } from '@schemas/categorization'
-
-export enum CategorizationType {
-  AUTO = 'AUTO',
-  ASK_FROM_SUGGESTIONS = 'ASK_FROM_SUGGESTIONS',
-  MEALS = 'MEALS',
-  BUSINESS_TRAVEL_TRANSPORTATION = 'BUSINESS_TRAVEL_TRANSPORTATION',
-}
-
-export interface AutoCategorization {
-  type: CategorizationType.AUTO
-  category: ApiCategorization
-}
-export interface SuggestedCategorization {
-  type: CategorizationType
-  suggestions: ApiCategorization[]
-}
-export type Categorization = AutoCategorization | SuggestedCategorization
-
-export type AccountIdentifierPayloadObject =
-  | {
-    type: 'StableName'
-    stable_name: string
-  }
-  | {
-    type: 'AccountId'
-    id: string
-  }
-  | {
-    type: 'Exclusion'
-    exclusion_type: string
-  }
+import { type NestedCategorization } from '@schemas/categorization'
 
 export const accountIdentifierIsForCategory = (accountIdentifier: AccountIdentifier, category: NestedCategorization): boolean => {
   if (accountIdentifier.type === 'AccountId') {
